--- components/pages/dashboard/Templates.tsx
@@ -1,17 +1,0 @@
-import { FolderOpen } from 'lucide-react';
-import React from 'react';
-
-const Templates = () => {
-  return (
-    <div className="text-center py-16">
-      <FolderOpen size={48} className="mx-auto text-muted-foreground/30 mb-6" />
-      <h2 className="text-xl font-semibold mb-3">Templates Coming Soon</h2>
-      <p className="text-muted-foreground">
-        We&apos;re working on bringing you a collection of cool designs to help
-        you get started.
-      </p>
-    </div>
-  );
-};
-
-export default Templates;
