+++ tsconfig.json
@@ -46,6 +46,7 @@
     ".next/types/**/*.ts"
   ],
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "*.test.*"
   ]
 }
